🤖 Refactor(client): alias 변경
