+++ apps/client/tsconfig.json
@@ -9,8 +9,7 @@
         "allowSyntheticDefaultImports": true, // default import 허용
         "baseUrl": ".",
         "paths": {
-            "@svgflow": ["src/svgflow/index.tsx"],
-            "@svgflow/*": ["src/svgflow/*"],
+            "@cloudflow": ["src/cloudflow/index.tsx"],
             "@cloudflow/*": ["src/cloudflow/*"],
             "@components/*": ["src/components/*"],
             "@hooks/*": ["src/hooks/*"],
