+++ packages/client/src/cloudflow/components/Edge/index.tsx
@@ -1,18 +1,48 @@
-import { Edge } from '@cloudflow/types';
 import { useTheme } from '@mui/material';
+import { useNodeContext } from '@cloudflow/contexts/NodeContext';
+import { Dimension, Edge } from '@cloudflow/types';
+import { calculateAnchorPoints } from '@cloudflow/utils';
+import { memo, useMemo } from 'react';
 
 type Props = {
     edge: Edge;
+    // visibleNodes: Node[];
+    dimension: Dimension;
 };
-export default ({ edge }: Props) => {
-    const { id, source, target } = edge;
+export default memo(({ edge, dimension }: Props) => {
+    const { id, sourceId, targetId, sourceAnchorType, targetAnchorType } = edge;
     const theme = useTheme();
     const color =
         theme.palette.mode === 'dark'
             ? theme.palette.grey[200]
             : theme.palette.grey[800];
 
-    const linePathD = `M ${source.point.x} ${source.point.y} L ${target.point.x} ${target.point.y}`;
+    const {
+        state: { nodes },
+    } = useNodeContext();
+    const [sourceNode, targetNode] = useMemo(() => {
+        const source = nodes.find((node) => node.id === sourceId);
+        const target = nodes.find((node) => node.id === targetId);
+        return [source, target];
+    }, [sourceId, targetId, nodes]);
+
+    //TODO: 보여지는 node에 대해서만 순회할지 고민.. 이렇게 하면 zoom/pan에서 너무많은 리렌더링이 발생함.
+    //
+    // const [sourceNode, targetNode] = useMemo(() => { const source = visibleNodes.find((node) => node.id === sourceId);
+    //     const target = visibleNodes.find((node) => node.id === targetId);
+    //     return [source, target];
+    // }, [sourceId, targetId, visibleNodes]);
+    //
+
+    if (!sourceNode || !targetNode) return null;
+
+    const sourceAnchors = calculateAnchorPoints(sourceNode.point, dimension);
+    const targetAnchors = calculateAnchorPoints(targetNode.point, dimension);
+
+    const sourcePoint = sourceAnchors[sourceAnchorType];
+    const targetPoint = targetAnchors[targetAnchorType];
+
+    const linePathD = `M ${sourcePoint.x} ${sourcePoint.y} L ${targetPoint.x} ${targetPoint.y}`;
 
     return (
         <g id={id}>
@@ -37,4 +67,4 @@ export default ({ edge }: Props) => {
             />
         </g>
     );
-};
+});
