+++ packages/client/src/cloudflow/components/Node/svgs/Server.tsx
