+++ packages/client/src/cloudflow/utils/index.ts
@@ -4,27 +4,7 @@ import {
     GRID_3D_WIDTH_SIZE,
     GRID_SIZE,
 } from '@cloudflow/constants';
-import { AnchorType, Dimension, Point, ViewBox } from '@cloudflow/types';
-import { RefObject } from 'react';
-
-export const getRelativeCoordinatesForViewBox = (
-    clientX: number,
-    clientY: number,
-    ref: RefObject<HTMLElement>,
-    viewBox: ViewBox
-): { x: number; y: number } => {
-    if (!ref.current) return { x: 0, y: 0 };
-
-    const zoomPan = ref.current.getBoundingClientRect();
-    const relativeX = clientX - zoomPan.left;
-    const relativeY = clientY - zoomPan.top;
-
-    const { point } = viewBox;
-    const x = (relativeX / zoomPan.width) * viewBox.width + point.x;
-    const y = (relativeY / zoomPan.height) * viewBox.height + point.y;
-
-    return { x, y };
-};
+import { AnchorType, Dimension, Point } from '@cloudflow/types';
 
 export const getDistance = (point1: Point, point2: Point) => {
     return Math.sqrt((point1.x - point2.x) ** 2 + (point1.y - point2.y) ** 2);
@@ -35,8 +15,7 @@ export const getSvgPoint = (flow: SVGSVGElement, cursorPoint: Point) => {
     svgPoint.x = cursorPoint.x;
     svgPoint.y = cursorPoint.y;
     const screenCTM = flow.getScreenCTM();
-    if (!screenCTM) return null;
-    return svgPoint.matrixTransform(screenCTM.inverse());
+    return svgPoint.matrixTransform(screenCTM!.inverse());
 };
 
 export const getNodeSizeForDimension = (dimension: Dimension) => {
@@ -49,26 +28,15 @@ export const getNodeSizeForDimension = (dimension: Dimension) => {
     return { width, height };
 };
 
-export const calculateAnchorPoint = (
-    type: AnchorType,
+export const calculateAnchorPoints = (
+    point: Point,
     dimension: Dimension
-): Point => {
-    const width = dimension === '2d' ? GRID_SIZE : GRID_3D_WIDTH_SIZE;
-    const height =
-        dimension === '2d'
-            ? GRID_SIZE
-            : GRID_3D_HEIGHT_SIZE + GRID_3D_DEPTH_SIZE;
-
-    switch (type) {
-        case 'top':
-            return { x: width / 2, y: 0 };
-        case 'right':
-            return { x: width, y: height / 2 };
-        case 'bottom':
-            return { x: width / 2, y: height };
-        case 'left':
-            return { x: 0, y: height / 2 };
-        default:
-            return { x: 0, y: 0 };
-    }
+): Record<AnchorType, Point> => {
+    const { width, height } = getNodeSizeForDimension(dimension);
+    return {
+        top: { x: point.x + width / 2, y: point.y },
+        right: { x: point.x + width, y: point.y + height / 2 },
+        bottom: { x: point.x + width / 2, y: point.y + height },
+        left: { x: point.x, y: point.y + height / 2 },
+    };
 };
