--- packages/client/src/svgflow/constants/index.ts
@@ -1,4 +1,0 @@
-export const GRID_SIZE = 90;
-export const GRID_3D_WIDTH_SIZE = 128;
-export const GRID_3D_HEIGHT_SIZE = 74;
-export const GRID_3D_DEPTH_SIZE = 37;
