--- packages/client/src/svgflow/contexts/EdgeContext/index.tsx
@@ -1,25 +1,0 @@
-import {
-    edgeReducer,
-    EdgeState,
-    initialState,
-} from '@svgflow/contexts/EdgeContext/reducer';
-import { createContext, useContext, useReducer } from 'react';
-
-const EdgeContext = createContext<{
-    state: EdgeState;
-    dispatch: React.Dispatch<any>;
-}>({
-    state: initialState,
-    dispatch: () => null,
-});
-
-export const EdgeProvider = ({ children }: { children: React.ReactNode }) => {
-    const [state, dispatch] = useReducer(edgeReducer, initialState);
-    return (
-        <EdgeContext.Provider value={{ state, dispatch }}>
-            {children}
-        </EdgeContext.Provider>
-    );
-};
-
-export const useEdgeContext = () => useContext(EdgeContext);
